Renamed files for `JoltJointImpl3D`
