+++ src/joints/jolt_hinge_joint_impl_3d.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "joints/jolt_joint_3d.hpp"
+#include "joints/jolt_joint_impl_3d.hpp"
 
 class JoltHingeJointImpl3D final : public JoltJointImpl3D {
 public:
