+++ src/spaces/jolt_space_3d.cpp
@@ -1,6 +1,6 @@
 #include "jolt_space_3d.hpp"
 
-#include "joints/jolt_joint_3d.hpp"
+#include "joints/jolt_joint_impl_3d.hpp"
 #include "objects/jolt_area_3d.hpp"
 #include "objects/jolt_body_3d.hpp"
 #include "servers/jolt_project_settings.hpp"
